Smaller refactorings and fixes
